* First attempt at getting feature branches to work with multisite setups.
* Don't need to import everything from FeatureBranches.py, actually.
* Woops, succeeded isn't a method I can call.
* Use branch, not build.
* Adding some debug info to FB teardown mapping function.
* Handle teardown removal of vhost a bit better if that vhost isn't found.
* Add a line to featurebranch vhosts which will be necessary later on when tearing down multisite feature branch builds.
* Revert "Add a line to featurebranch vhosts which will be necessary later on when tearing down multisite feature branch builds."
This reverts commit db1b1aabe4770d0a7e1179cc27743f6a9e37f5e5.
* Don't need to do any silly vhost stuff, just need to sort the result from find and grep when tearing down a site.
* Bit of debug text in fabfile-teardown.py.
* Use sort -r
* Need to escape the square brackets in grep command when configuring teardown mapping.
* Scrap using the sed() fabric function altogether when adjusting settings.php.
* Paths to multisite sites are wrong when preparing the database during a feature branch build.
* Removing the use of Fabric's sed() command when reverting settings.php.
* Trying to use drush wrapper with some prepare_database() stuff.
* First attempt at ensuring all multisite sites are reverted should a failure occur during the build.
* Cron generation is broken for multisite builds, so fixing that.
* Add back the stable_build variable which is needed in the _revert_settings() function. Woops.
* Also need the settings_file variable. Ugh.
* Making the print statements in _revert_settings() a bit more helpful.
* Reworking how we check if a site exists, as the current method means new additions to an already existing multisite won't actually be deployed.
* Sites are in the sites directory...ugh.
* Need to wrap the site check in a warn_only=True, otherwise the build will abort if a new site is added to a multisite install.
* Trying to re-organise when db backups happen, when sites go offline and online, and when the live symlink is adjusted.
* Only run the go_online() task when do_updates is True.
* Applying some stricter logic to certain aspects of the deployment process.
* sort -r doesn't work when tearing down sites, so we need to use sort --version-sort instead.
* Spelling errors. Sigh.
* Making app IPs non-optional.
* Blimmin blimmin blimmin blimmin
* Adding a fabfile-editorial.py script which clones down the repo, adjusts config (settings.php, files symlink) and adjusts live symlink on an editorial site.
* Adding a comment block to the fabfile-editorial.py script explaining what it is used for.
* Need to import the Drupal module in the fabfile-editorial.py script.
* r33317 support the use of an .env file instead of a parameters.yml in Symfony. Default to using parameters.yml, but allow either or both methods (e.g whilst an .env technique is being phased in on a project, they may both need to exist)
* Disable StrictHostKeyChecking and UserKnownHostsFile when rsyncing files up to a destination server in the fabfile-sftp.py script.
* Account for parameters being passed in as strings by Jenkins
* Clutching at straws with this Drush db creation stuff. Will waiting a bit help, I wonder.
* Need to import the time module in InitialBuild.py.
* Fix where the .env.dist file is copied from if an environment-specific version was not found
* Use the -l option when creating a database dump for a feature branch build.
* Use the -l option when creating a database dump for a feature branch build.
* In prepare_database, dump_file is incorrectly set to None when the source db is on a different server.
* Need to adjust live symlink on all hosts during a Magento build.
* S3 upload needs 'branch' not 'buildtype' for path.

+++ common/BuildTeardown.py
@@ -6,21 +6,22 @@ import os
 
 @task
 def remove_vhost(repo, branch, webserver, alias):
-  if alias is None:
-    alias = repo
   with settings(warn_only=True):
     print "===> Unlinking and removing %s vhost..." % webserver
     # We grep the config files for the correct symlink to be sure we delete the right one
-    conf_file = sudo("find /etc/%s/sites-enabled/ -name '*%s*' -print0 | xargs -r -0 grep -H 'live.%s.%s' | head -1 | awk '{print $1}' | cut -d '/' -f 5 | cut -d ':' -f 1" % (webserver, alias, repo, branch))
+    conf_file = sudo("find /etc/%s/sites-enabled/ -name '*%s*' -print0 | xargs -r -0 grep -H 'live.%s.%s' | sort --version-sort | head -1 | awk '{print $1}' | cut -d '/' -f 5 | cut -d ':' -f 1" % (webserver, alias, repo, branch))
 
     print "%s conf file is: %s" % (webserver, conf_file)
-    sudo("unlink /etc/%s/sites-enabled/%s" % (webserver, conf_file))
-    sudo("rm /etc/%s/sites-available/%s" % (webserver, conf_file))
+
+    if sudo("stat /etc/%s/sites-available/%s" % (webserver, conf_file)).failed:
+      print "===> Could not find a vhost! Don't try to remove anything this time."
+    else:
+      sudo("unlink /etc/%s/sites-enabled/%s" % (webserver, conf_file))
+      sudo("rm /etc/%s/sites-available/%s" % (webserver, conf_file))
+
 
 @task
-def remove_http_auth(repo, branch, webserver, alias=None):
-  if alias is None:
-    alias = repo
+def remove_http_auth(repo, branch, webserver, alias):
   print "=> Removing htpasswd file, if it exists..."
   with settings(warn_only=True):
     if sudo("stat /etc/%s/passwords/%s.%s.htpasswd" % (webserver, alias, branch)).failed:
@@ -28,10 +29,18 @@ def remove_http_auth(repo, branch, webserver, alias=None):
     else:
       sudo("rm /etc/%s/passwords/%s.%s.htpasswd" % (webserver, alias, branch))
 
+
 @task
-def remove_cron(repo, branch, alias=None):
-  if alias is None:
-    alias = repo
+def remove_cron(repo, branch, alias):
   with settings(warn_only=True):
     print "===> Removing cron file..."
     sudo("rm /etc/cron.d/%s_%s_cron" % (alias, branch))
+
+
+@task
+def remove_repo_code(repo, branch):
+  with settings(warn_only=True):
+    # Remove site directories
+    print "===> Unlinking live symlink and removing site directories..."
+    sudo("unlink /var/www/live.%s.%s" % (repo, branch))
+    sudo("rm -rf /var/www/%s_%s_*" % (repo, branch))
